contestIdInputコンポーネント切り分け
